requirements.txt
requests
litserve
llama-index-core
llama-index-llms-ollama
llama-index-vector-stores-qdrant
llama-index-embeddings-fastembed
llama-index-readers-json